ADD: exists function

+++ cam_mammogram.py
@@ -27,7 +27,7 @@ from skimage.util import view_as_windows
 from tensorflow.keras.models import save_model#, load_model
 # from tensorflow.keras.metrics import F1Score
 from tensorflow.keras.callbacks import EarlyStopping
-
+from os.path import exists
 # from pandas import DataFrame
 # from itertools import chain
 """
@@ -724,42 +724,44 @@ def main():
     patch_architecture_dense = create_patch_model_dense((GLOBAL_X,GLOBAL_Y,3))
     plt.figure(figsize=(15, 8))
     previous_val_acc = 0
-    early_stopping = EarlyStopping(monitor='val_accuracy', patience=3, restore_best_weights=True)
+    early_stopping = EarlyStopping(monitor='val_loss', patience=5, restore_best_weights=True)
 
     print('Train denseNet')
-    for i in range(1):
-        history = patch_architecture_dense.fit(X_train, y_train, epochs=50, batch_size=64,callbacks=[early_stopping],
+    if not exists('patch_DenseNet.h5'):
+
+        for i in range(1):
+            history = patch_architecture_dense.fit(X_train, y_train, epochs=50, batch_size=64,callbacks=[early_stopping],
                                         validation_data=(x_val, y_val), verbose=1)
 
-        plt.subplot(1, 2, 1)  # 1 row, 2 columns, 1st subplot
-        plt.plot(history.history['accuracy'], label=f'Training Accuracy ({i}th iteration)')
-        plt.plot(history.history['val_accuracy'], label=f'Validation Accuracy ({i}th iteration)')
-        #plt.plot(history.history['val_f1_score'], label=f'Validation F1 Score ({i}th iteration)')
-        plt.title('DenseNet Baseline Model Accuracy History')
-        plt.xlabel('Epoch')
-        plt.ylabel('Accuracy')
-        plt.legend()
+            plt.subplot(1, 2, 1)  # 1 row, 2 columns, 1st subplot
+            plt.plot(history.history['accuracy'], label=f'Training Accuracy ({i}th iteration)')
+            plt.plot(history.history['val_accuracy'], label=f'Validation Accuracy ({i}th iteration)')
+            #plt.plot(history.history['val_f1_score'], label=f'Validation F1 Score ({i}th iteration)')
+            plt.title('DenseNet Baseline Model Accuracy History')
+            plt.xlabel('Epoch')
+            plt.ylabel('Accuracy')
+            plt.legend()
 
-        plt.subplot(1, 2, 2)  # 1 row, 2 columns, 2nd subplot
-        plt.plot(history.history['loss'], label=f'Training Loss ({i} iteration)')
-        plt.plot(history.history['val_loss'], label=f'Validation Loss ({i} iteration)')
-        plt.title('Densenet Baseline Model Loss History')
-        plt.xlabel('Epoch')
-        plt.ylabel('Loss')
-        plt.legend()
-        if history.history['val_accuracy'][-1] > previous_val_acc:
-            save_patch_model = patch_architecture_dense
-            print(f'({i} iteration) best model: {history.history["val_accuracy"][-1]}')
+            plt.subplot(1, 2, 2)  # 1 row, 2 columns, 2nd subplot
+            plt.plot(history.history['loss'], label=f'Training Loss ({i} iteration)')
+            plt.plot(history.history['val_loss'], label=f'Validation Loss ({i} iteration)')
+            plt.title('Densenet Baseline Model Loss History')
+            plt.xlabel('Epoch')
+            plt.ylabel('Loss')
+            plt.legend()
+            if history.history['val_accuracy'][-1] > previous_val_acc:
+                save_patch_model = patch_architecture_dense
+                print(f'({i} iteration) best model: {history.history["val_accuracy"][-1]}')
 
-    plt.tight_layout()  # Adjust subplot spacing for better appearance
-    plt.savefig('training_accuracy_loss_DenseNet.png', dpi=400)
-    test_results = patch_architecture_dense.evaluate(X_test, y_test)
-    print(f'Test Accuracy: {test_results[1]}')
-    print(f'Test Loss: {test_results[0]}')
-    with open('test_results_dense.txt', 'w') as file:
-        file.write(f'Test Accuracy: {test_results[1]}\n')
-        file.write(f'Test Loss: {test_results[0]}\n')
-    save_model(save_patch_model,'patch_DenseNet121.h5')
+        plt.tight_layout()  # Adjust subplot spacing for better appearance
+        plt.savefig('training_accuracy_loss_DenseNet.png', dpi=400)
+        test_results = patch_architecture_dense.evaluate(X_test, y_test)
+        print(f'Test Accuracy: {test_results[1]}')
+        print(f'Test Loss: {test_results[0]}')
+        with open('test_results_dense.txt', 'w') as file:
+            file.write(f'Test Accuracy: {test_results[1]}\n')
+            file.write(f'Test Loss: {test_results[0]}\n')
+        save_model(save_patch_model,'patch_DenseNet121.h5')
 
     print('Train ResNet')
     previous_val_acc = 0
